Add helpful validation to `run()`
When forgetting to provide the context as first argument, the method
gets confused and would run a command with the first argument missing.

With this change this cannot happen anymore.

On the downside, this requires weird mocking for tests, and could be
avoided altogether by storing the context in the command itself.