(PDK-508) Setup easy testing of test_module
By pulling in puppetlabs_spec_helper into the main project, we can run
tests for the `test_module` in the main test suite without friction.
This makes it easier and quicker to run this kind of tests.

See the next commits for usage.

This requires https://github.com/puppetlabs/puppetlabs_spec_helper/pull/234