(PDK-508) implement autorequire and friends
The spec/classes/autorequire_cycle_spec.rb test shows that the
autorelations implementation of the Resource API does build relations.
Since rspec-puppet doesn't provide matching on compile errors, this
builds a test case that compiles when make_cycle is false, thereby
proving that there is no error in the manifest, but fails with a
dependency cycle error when make_cycle is set to true.