(PDK-882) Allow `puppet resource` to work on non-existing resources
When `puppet resource` tries to find a specific resource, it only passes
a hash with the title through. Puppet then calls validate, which
fails, because the `ensure` property is mandatory.

This change ignores a missing `:ensure` value if a call from `puppet resource`
is detected.

Added a bunch of acceptance-level tests to make sure that the `apply`
case still works.