(PDK-895) Basic Array support
Use the new `mungify` function to replace the entire validate/munge
cycle on Parameters and Properties.

This includes a couple of hacks to not break other parts of puppet.

Also includes some fixes were test data for the patterns did not actually
match the expectations (either Integers instead of Strings, or only seven
digits, instead of eight).