Rename internal validation state detection
It turns out that a new type instance with a Hash as argument is only
used by the RAL's `find` method, if there is no matching instance returned
from `instances`. While this is the method used by the `puppet resource`
CLI, it is not indicative of being called from it (when the instance
exists, this code path is not hit). This commit changes the name of the
variable to better represent what it actually signifies.