(PDK-1007) Extract current state fetching and caching
With the last few changes the rsapi would call into get multiple times
for a type during a single run. This commit now pulls the retrieval of
the current state into `refresh_current_state` and leaves the caching
vs calling decision up to the caller.

As a side-effect the strict get-returns-canonical-values check now runs
at a time when puppet is not aborting the transaction for errors, and
the "exercising a device provider using `puppet resource` with strict
checking at error level deals with canonicalized resources correctly"
test case does not produce a non-zero exit code anymore.

Since now the state of the resource is correctly cached, the "exercising
simple_get_filter" test needed to be adjusted to use noop to display the
test result value.