(PDK-1007) return full `Puppet::Type`s from instances
To allow the `resources` type to purge resource instances, the
`instances` method needs to return full `Puppet::Type` object
instances. This commit changes the Resource API's definition
of the `instances` method to do so, and fix up all adjacent
issues.

As of yet, this code still fails tests because it is not passing
on the "current" values to the `Puppet::Type` instances returned.
Instead `get` gets called through retrieve for each individual
instance, causing both performance issues and test failures.