(PUP-9747) Relax validation for bolt
Bolt always provides all computed connection info attributes to tasks,
instead of only the ones the user provided. Therefore we can't be strict
when validating the passed in arguments.

To limit the impact to bolt users, this change makes the relaxed validation
trigger only when a `'remote-transport'` key is specified, as that is
always provided by bolt.

This also contains the docs changes from https://github.com/puppetlabs/puppet-specifications/pull/142