(FM-8553) Remove all caching from list_all_transports
It turned out that the environment object in the puppetserver does not survive for longer than a request, making the ObjectIdCacheAdapter creating a new cache everytime. Since the autoloader was using `require` to load schemas, this lead to the cache not being refresehd, and the API would return wrong (empty) results.

This change fixes this by not caching at all and using `load` instead of `require`, so that the schemas are always freshly loaded.