(maint) test handling of untyped and single-element arrays
This is in response to PUP-10428. The tests are reproducing the conditions
reported.

The integration test does show the mishandling of array values (rsapi is
not using `array_matching`). The acceptance tests do show correct messages and changes
being effected despite this for arrays with multiple and single elements,
using the Array type and not using it.

While I do not fully understand (due to limited time for analysis) what
resource api is doing differently to not trigger the issue reported, the
fact that the enhanced acceptance tests are passing gives me confidence
to pass on further investigative work here.