(GH-242) Ensure catalog passing when raw_catalog_access flag specified
This commit updates the BaseContext class initialization method
to take a second argument, catalog (default to nil), which
ensures that the @catalog instance variable is populated with
the value passed.

It also updates the context method in resource_api to pass the
catalog to that initializer if specified, or pass nil if not.

This does not add unit or acceptance tests for the behavior.