(GH-225) Refactor insync? to return result and change message
This commit modifies the def_custom_insync? handling in the property to
expect that calling insync? against the provider will return both the
provider_insync_result and a change_message, defaulting to nil for both
if nothing is returned.

If only one value is returned, it is treated as the result and the message
is nil.

If the result is nil or the change message is nil or empty, change_to_s
is not modified and reports the output of Puppet::Property.change_to_s as
the change log if needed.

If the result is not nil and the change message has a value, the change
message is used for the change log if needed.

The method also does some guardrail handling for result values:

- If the result is nil, fall back on Puppet::Property.insync? to check
  whether the property is in the desired state or not.
  - If the result is nil and the attribute being checked is ensure,
    validate the comparison by comparing the values as strings.
- If the result is a real boolean, return that
- If the result is anything else, raise an explanatory Puppet::DevError.