Generate changelog without `v` prefix
This repo historically used v1.x.x tags, but we stopped tagging with the `v`
prefix in 1.8.2. However, when doing the 1.8.14 release, both 1.8.14 and v1.8.14
tags were created. So this commit excludes the v1.8.14 tag to avoid creating an
empty changelog diff for `v1.8.14...1.8.14`

And it updates the Rakefile to not prefix the future tag with `v`.