(PA-5641) Update Puppet and Ruby versions in rspec
This commit updates the rspec GitHub Actions workflow to drop Puppet 6
and Ruby 2.5 and add Puppet 8 and its corresponding Rubies (MRI 3.2 and
JRuby 9.4.2.0).

Because of Ruby incompatibilities, this commit also updates tests to
run against specific pairings of Puppet and Ruby, instead of testing
the entire matrix against everything else.