(PA-5641) Correct method to use optional argument
Prior to this commit, the `processing` method's final argument,
`message`, used a colon, which is the syntax for a keyword argument.
This does not seem to be the intended behavior given how the method is
called in tests, and may be an artifact from earlier versions of Ruby.

This commit updates the `processing` method's final argument from a
keyword argument to an optional positional argument by changing the
colon to an equals sign.