(PA-5641) Update tests with options hashes
Ruby >= 3.0 is more strict about interpreting keyword arguments than
previous Rubies. This commit updates rspec tests to receive an options
hashes instead of keyword arguments to enable these tests to run on Ruby
>= 3.0.

Additionally, this commit disables the Style/BracesAroundHashParameters
cop in Rubocop, as it is a style choice that conflicts with the
functionality of these options hashes. This cop may also be generating
false positives because the version of Rubocop that we're running is
over five years old and was released before the separation of positional
and keyword arguments that appeared in more recent versions of Ruby[1].

[1] https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/