(PA-4639) Migrate away from AppVeyor
This commit migrates CI for Windows from AppVeyor to GitHub Actions.

When testing against Puppet 8 on Ruby 3.2 on Windows, we manually
install libffi through a new "integration" group in the Gemfile. This
is done for a few reasons:

- Ruby 3.2 no longer bundles libffi, as previous versions of Ruby had.
- Tests on Windows require libffi.
- While the Windows-specific Puppet gem lists libffi as a dependency,
  the C library used to build that gem (MinGW) differs from the C
environment set up on GitHub runners (ucrt).

For more information, see PA-5406