(PA-5641) Change github bump action and release
The old github action from anothrNick/github-tag-action only used
lightweight tags; this caused compatibility problems with the way
puppetlabs-jenkins used annotated tags. In order to keep the tags
consistent, this commit uses mathieudutour/github-tag-action
which supports annotated tags. Since the previous attempt to bump
skipped the pre-bumped version, that old tag is still available
and can be used to synchronize the version in code and in tag.