Fix CODEOWNER regex to include devx and phoenix
When PRs are raised the CODEOWNERS.md is assessed via regex to determine the allowed approvers. The     CODEOWNERS file uses the last matching pattern for each file. In this case, there are two patterns t    hat match all files, and the last one is "@puppetlabs/phoenix". This is why only the Phoenix team is     being requested for review.  To have both teams requested for review, the regex needs to be in one     line like this: "@puppetlabs/devx @puppetlabs/phoenix". This way, both teams will be automatically r    equested for review when a pull request is opened.

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>