Fix rubocop warnings.
All of these warnings were fixed in order to correct a rubocop syntax error occuring in jruby but not ruby.  After pinning jruby's (9.3 and 9.4) to an older version of rubocop than ruby's versions (2.7 and 3.2),  the older rubocop did not recognize some newer cops listed in the `rubocop_todo.yml`.  The solution was to auto-correct just these new exceptions.  This is a temporary solution, however, as newer cop warnings may cause the same issue again.

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>