Fix rubocop errors
Recent rubocop updates broke the resource_api's CI:

```ruby
➜  puppet-resource_api git:(main) bundle exec rubocop
Error: Property AutoCorrect of cop FactoryBot/CreateList is supposed to be a boolean and contextual is not.
+ set +x
➜  puppet-resource_api git:(main)
```

Although the fix for other repositories was to upgrade rubocop, an upgrade was not possible on this repo because of the `jruby` and `ruby` engine testing.  The solution was to pin the rubocop dependencies to the last working nightly versions:

```ruby
gem 'rubocop', '~> 1.48.1', require: false
gem 'rubocop-rspec', '~> 2.20.0', require: false
gem 'rubocop-performance', '~> 1.17.1', require: false
```

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>